+++ app/services/index.ts
@@ -1,2 +1,1 @@
 export * from "./authService";
-export * from "./userService";
